Move time to io.spine.base.
