+++ base/src/main/java/io/spine/base/Time.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.time;
+package io.spine.base;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.protobuf.Timestamp;
