+++ base/src/test/java/io/spine/validate/MessageValidatorShould.java
@@ -79,7 +79,7 @@ import static com.google.protobuf.util.Timestamps.subtract;
 import static io.spine.Identifier.newUuid;
 import static io.spine.protobuf.TypeConverter.toMessage;
 import static io.spine.test.Verify.assertSize;
-import static io.spine.time.Time.getCurrentTime;
+import static io.spine.base.Time.getCurrentTime;
 import static java.lang.String.format;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
